♻(ingestion-presentation) isolate concours views
