+++ src/web/presentation/ingestion/urls.py
@@ -2,10 +2,10 @@ from django.urls import path
 
 from presentation.ingestion.views import (
     ArchiveOffersView,
-    ConcoursUploadView,
     OffersListView,
     OffersUpsertView,
 )
+from presentation.ingestion.views.concours import ConcoursUploadView
 from presentation.ingestion.views.metiers import MetiersListView
 from presentation.ingestion.views.sources import SourcesListView
